requirements.txt
pyTelegramBotAPI==4.14.0
python-dotenv==1.0.0
fpdf2==2.7.6
requests==2.31.0
aiohttp==3.9.1
